detect cycles in directed graphs
determine whether or not a graph contains a cycle. A cycle in a directed graph is a set of nodes {x1, …, xn} in which an edge exists from xi to xi+1 for all 1 <= i < n and an edge from xn to x1

By using unordered_map to keep track of each nodes.